function/requirements.txt
azure-functions
azure-cosmos
azure-storage-file-datalake